Fix so tests don't poop temporary working directories all over the source tree
- test_harness now creates a clean temporary directory at the beginning of each test.
Previously, on the compile function would create it on demand, and other code depended
that having already been created.
- Fixed a bunch of places that had hard-coded working directory paths to use a common
variable defined in test_harness.py.

There is now just one working directory at the top level of the tests source
directory.  It might be preferable to move it out of the test directory entirely.
However, I resisted just using tempfile, because, for debugging, it is often
useful to be able to get at the intermediate files.