Implement write support in SDMMC and fix a bug
- I may have figured out why reads seemed to fail on FPGA, always returning
0xff. I overlooked that the data transfer is supposed to be preceeded by a
'DATA_TOKEN', which I wasn't waiting for. I haven't tested this on FPGA yet.
- To get write support working in sim_sdmmv.sv, I modified it so it doesn't
load the entire block device contents into memory, but reads and writes blocks
on demand as needed. This has the nice side effect of removing hard-coded
size limits. Because verilator doesn't support $fwrite/$fseek, I needed to
use $c() to emit code for that. I added code for other simulators, but don't
have an easy way to test it, so it may not work correctly.
- I made a similar change to the emulator sdmmc implementation. It used
to mmap the device contents (which would work just fine for writing), but
I now read and write blocks on demand so it's similar to the hardware implementation.
- Renamed identifiers to be clearer