Add ability to input serial characters into emulator
When a character is written to the UART when running on the emulator, it
is directed to stdout for the emualtor process. This change makes it so
any characters that come into stdin for the emulator appear to be
transfered into the UART.

This allows interactive terminal sessions on the emulator command line,
useful for bringing up a kernel with a shell, which is my next project.