Transition to using runtime library included with compiler
As part of this, I needed to update the kernel to (partially) support PIC, by
loading the gp register. This is a hack because I haven't implemented support
in the compiler for loading gp yet.

compiler-rt is built with PIC by default, since the compiler makefiles assume it
could be statically linked into a shared object.

I also added a 'abort' call, since compiler-rt expects that to be present.