Align verilog SDMMC simulator with emulator
I've tossed around the idea of using the emulator C implementation from the
hardware model using PLI, which would be fairly straightfoward at this point,
but it's a bit ugly from a build perspective. For now, keeping these files
largely in sync in terms of identifiers and structure so it's easy to add
new features to both. I'm doing a bunch of restructuring to prepare to
add support for a few new commands needed for larger cards, including
the CMD8, which uses the R7 response type.