Update to new version of compiler toolchain (SVN 329504)
This upgrade found a bug in libos. dlmalloc expects sbrk to zero the memory
it returns by default (based on MMAP_CLEARS define). sbrk was not doing this.
This would cause calloc to sometimes return non-zeroed memory. The new
compiler is apparently smart enough to see that a pointer was allocated
with calloc, and will skip initializing members that are zero. This broke
tests/whole-program/fannkuch.

Changed docker scripts because LLVM changed the compiler naming convention
to remove the minor version.

I'm not sure how the btree test ever worked: it put constant string pointers
into each node, then was trying to free them. With the new compiler version,
this was tripping an assertion. dlfree checks that the previous block is in use,
which looks at a bit pattern in the word before the pointer. A change in program
layout could cause it to be pointing at a different value.