Fix failing test caused by new toolchain
The challenge with floating point values in tests is that the value can
change slightly based on the order of operations, which can legally
be affected by optimizations. The results aren't visible (often just
1 ULP), but the tests validate exact bit patterns.

This test would be more robust if it did a comparison with the target image
and allowed some small amount of variability. Currently it computes a SHA
checksum of the framebuffer.

Updated the checksum for the new toolchain.