Make path configuration in build system more flexible
The path to the build compiler can now be overridden with a cached cmake variable like this:

    cmake -DNYUZI_COMPILER_ROOT=/path/to/compiler .

I've also made modifications to hardcoded paths in the test project to use this. The test
project now loads paths out of tests.conf rather than hardcoding or trying to guess them.
This is stored in-tree, which is not usual, but the only way for the test harness to locate
it consistently.

Previously, there was some code in the test harness that searched a few well known locations
for build outputs. Now it is much more flexible and can work with an out-of-tree build that
is put anywhere.

As part of this, I also moved the working directory of the tests (where it puts intermediate
generated files) in the build directory. I don't know if this is the best place for it, since
it's not strictly part of the build, but I think it's better than leaving them in-tree.