Fix build breakages introduced by last two changes
This probably also addresses issue #161. The issue is that there should be a dependency:

sceneview -> sceneview_fsimage -> resource_bin

The second target is created automatically by the add_nyuzi_executable macro. Since
the CMakeLists.txt for sceneview didn't know about that target, it didn't add a
dependency between it and resource_bin. When performing parallel make (which CI
does), often the fsimage would be built before resource_bin.

I've also added a DEPENDS in the custom command to make the filesystem image on the
source files. However, this alone doesn't fix the problem, because CMake doesn't
appear to use that as part of the dependency chain. My solution was to call add_dependency
from sceneview's CMakeList.txt between sceneview_fsimage and resource_bin.

Also noticed that make failed silently and the error was only caught later. I added set -e
to the top of the CI script to halt immediately in this case.

This also fixes the LLDB tests, where one of the variables wasn't updated from a previous change.
This doesn't run during CI (because LLDB isn't installed in the build container), so I didn't
see it immediately.