emulator: Fix minor issue with poll_interrupts
The code was treating the return value of can_read_file_descriptor as
an integer instead of a boolean. This didn't cause any functional
issues: although the clause that attempts to detect an error from
can_read_file_descriptor will never be hit, in an error condition,
it will still return true, then the subsequent read will detect an
error.