Update fields in AXI bus interface
Table A10-1 and A10-2 in the AXI spec (rev D) indicates which signals are
required and optional, and default values for optional signals. I was
hardwiring several values to the default value anyway, so I just removed them.
There were a few that were marked required that I didn't have, so I added them
for completeness.

This was mostly done to clean up the design and remove unused cruft, but one
other advantage is that the minimal required subset of AXI-4 signals is close
to AXI-4 lite, so I can cleanly use the same interface for both rather than
creating a separate interface for the latter.