Add basic perf test and fix issue
The store count was too large, because it would count when a store was requested
instead of when it was actually sent to the L2 cache. If there was a rollback
because the store queue was full, this would be counted as an extra store. To
address this, I now generate the store perf event in the store queue.