Fix builds (take 2)
I had made what I thought was an innocuous "cleanup" to the code to remove
a priority encoder, which turned out to be necessary. As it turns out, the priority
order needed to match the order of the block above it, which was subtle and
error prone. I combined the blocks to hopefully make the logic clearer. I also
added an assertion to check for the specific bug I introduced.