Issue #139: Remove test_harness.TimedProcessRunner
I also removed 'kill_gently' in favor of manually restoring the termios
state (the emulator changes the terminal settings to disable local echo,
so it wouldn't work correcterly adter timeouts). As it turns out, I didn't
need to do that, because the process is not killed after a timeout, but it's
more robust anyway, so I left it.