Use PIL for image conversion instead of Imagemagick
Previously code that needed to read or write image files spawned
ImageMagic convert as a subprocess, which was clunky. Now they
use the python 'PIL' library to do it directly.
This adds an extra step to setup, but is a bit easier now since
MacOS installs pip3 by default and Linux can install it using
apt.