Workaround for issue #183: Synthesized design doesn't boot on FPGA
Use elf2hex to produce the binary instead of trying to make the linker
emit a raw binary. This wastes 4k of ROM because of the way the linker
aligns segments, but it still fits.

I have not tested this on hardware or otherwise, so this is a speculative
fix.