Attempt to fix #183 (doesn't boot on FPGA), take 2
Explicitly specify the .rodata and .eh_frame sections in the linker script.
Previously, it was emitting these first, so when the CPU jumped to the lowest
address in the bootrom, it would begin executing garbage.  Now, the first
instruction of text (_start) is the first word emitted.

I tested this two ways:

1. Put  a .long directive with 0xdeadbeef at the beginning of start.S
and ensuring that word showed up first in the generated .hex file.
2. Disassembling start.S with objdump and checked that the beginning hex
values matched the first few instructions.

I have not yet tested this on FPGA.